Start code sharing between WebsiteMode and ShareMode
